// packages/api/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserStatus = {
  ONLINE: 'ONLINE',
  AWAY: 'AWAY',
  OFFLINE: 'OFFLINE'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]


export const MemberRole = {
  OWNER: 'OWNER',
  ADMIN: 'ADMIN',
  MEMBER: 'MEMBER'
} as const

export type MemberRole = (typeof MemberRole)[keyof typeof MemberRole]


export const ChannelType = {
  PUBLIC: 'PUBLIC',
  PRIVATE: 'PRIVATE',
  DM: 'DM'
} as const

export type ChannelType = (typeof ChannelType)[keyof typeof ChannelType]


export const MessageType = {
  TEXT: 'TEXT',
  SYSTEM: 'SYSTEM',
  FILE: 'FILE'
} as const

export type MessageType = (typeof MessageType)[keyof typeof MessageType]